delete conflict hook start with failing test
